requirements.txt
streamlit>=1.40.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
pandas>=2.2.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
Pillow>=10.0.0
